Adding EdDSA client C interface
